Add identity

+++ Tajan.Standard.Domain/Settings/RabbitMqSettings.cs
@@ -3,7 +3,7 @@
 public class RabbitMqSettings
 {
     public string Host { get; set; }
-    public int Port { get; set; }
+    public ushort Port { get; set; }
     public string Username { get; set; }
     public string Password { get; set; }
     public string VirtualHost { get; set; }
